Moved the calculation logic into one shared module
and fixed xp readme
